Port signal test.
Extended clock testbench for simple test of port and signal.
symsysc update with some bugfixes.
